+++ client/examples/soem.cpp
@@ -3,17 +3,18 @@
 // Created Date: 19/05/2020
 // Author: Shun Suzuki
 // -----
-// Last Modified: 25/12/2020
+// Last Modified: 30/03/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2020 Hapis Lab. All rights reserved.
 //
 
+#include "link/soem.hpp"
+
 #include <iostream>
 
 #include "autd3.hpp"
 #include "runner.hpp"
-#include "link/soem.hpp"
 
 using namespace std;